# [Perf] `ActorPath`: tried caching `Join()` output