Alternative to https://github.com/akkadotnet/akka.net/pull/5046

Ultimately, didn't seem to move the needle much - plus the memory footprint tradeoffs don't seem worth it to me for systems that have a very large number of actors.

```ini
BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19041.985 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.203
  [Host]     : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT
  DefaultJob : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT
```

|                                     Method |      Mean |     Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------------- |----------:|----------:|---------:|-------:|------:|------:|----------:|
|                            ActorPath_Parse | 565.45 ns | 10.378 ns | 9.707 ns | 0.1469 |     - |     - |     616 B |
|                           ActorPath_Concat |  57.81 ns |  1.216 ns | 2.065 ns | 0.0305 |     - |     - |     128 B |
|                           ActorPath_Equals |  20.21 ns |  0.322 ns | 0.285 ns |      - |     - |     - |         - |
|                         ActorPath_ToString | 129.42 ns |  1.091 ns | 1.021 ns | 0.0267 |     - |     - |     112 B |
|            ActorPath_ToSerializationFormat | 156.49 ns |  2.290 ns | 2.030 ns | 0.0267 |     - |     - |     112 B |
| ActorPath_ToSerializationFormatWithAddress | 175.53 ns |  1.431 ns | 1.338 ns | 0.0267 |     - |     - |     112 B |