@Aaronontheweb 2021-06-02
``` ini

BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19041.985 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.203
  [Host]     : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT
  DefaultJob : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT


```
|                                     Method |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------------- |----------:|----------:|----------:|-------:|------:|------:|----------:|
|                            ActorPath_Parse | 580.69 ns | 11.569 ns | 13.772 ns | 0.1392 |     - |     - |     584 B |
|                           ActorPath_Concat |  59.04 ns |  1.219 ns |  1.628 ns | 0.0267 |     - |     - |     112 B |
|                           ActorPath_Equals |  19.42 ns |  0.298 ns |  0.264 ns |      - |     - |     - |         - |
|                         ActorPath_ToString | 205.58 ns |  2.330 ns |  2.180 ns | 0.0401 |     - |     - |     168 B |
|            ActorPath_ToSerializationFormat | 227.92 ns |  2.169 ns |  1.694 ns | 0.0401 |     - |     - |     168 B |
| ActorPath_ToSerializationFormatWithAddress | 250.58 ns |  4.903 ns | 12.917 ns | 0.0401 |     - |     - |     168 B |